requirements.txt
accelerate==0.33.0
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.5
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pydeck==0.9.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
safetensors==0.4.3
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.37.0
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.43.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
wikipedia==1.4.0
